credy/requirements.txt
django-environ
django
djangorestframework
environ
requests
djangorestframework_simplejwt
response